requirements.txt
openai
python-dotenv
matplotlib
networkx
datetime